--- a/go.mod
+++ b/go.mod
@@ -4,9 +4,9 @@ go 1.13
 
 require (
 	github.com/datadog/czlib v0.0.0-20160811164712-4bc9a24e37f2
-	github.com/gogo/protobuf v1.3.1
-	github.com/paulmach/orb v0.1.6
+	github.com/paulmach/orb v0.1.3
 	github.com/paulmach/protoscan v0.2.1
 	golang.org/x/time v0.0.0-20190921001708-c4c64cad1fd0
-	google.golang.org/genproto v0.0.0-20200526211855-cb27e3aa2013 // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+	google.golang.org/protobuf v1.27.1
 )