--- a/internal/osmpb/generate.go
+++ b/internal/osmpb/generate.go
@@ -1,10 +1,3 @@
 package osmpb
 
-// gogo protobuf seems to be about 10% faster
-// https://github.com/gogo/protobuf
-//
-// setup:
-//    brew install protobuf
-//    go get github.com/gogo/protobuf/...
-//
-//go:generate protoc --proto_path=$GOPATH/src/:$GOPATH/src/github.com/gogo/protobuf/protobuf:. --gogofaster_out=. osm.proto
+//go:generate protoc  --go_opt=module=github.com/paulmach/osm/internal/osmpb  --go_out=. osm.proto