--- a/osmpbf/internal/osmpbf/generate.go
+++ b/osmpbf/internal/osmpbf/generate.go
@@ -1,5 +1,3 @@
 package osmpbf
 
-// gogo protobuf seems to be about 10% faster
-// https://github.com/gogo/protobuf
-//go:generate protoc --proto_path=$GOPATH/src/:$GOPATH/src/github.com/gogo/protobuf/protobuf:. --gogofaster_out=.  fileformat.proto osmformat.proto
+//go:generate protoc  --proto_path=. --go_opt=module=github.com/paulmach/osmpbf/internal/osmpbf  --go_out=.  fileformat.proto osmformat.proto