… go.sum
… internal/osmpb/osm.pb.go
… osmpbf/internal/osmpbf/fileformat.pb.go
… osmpbf/internal/osmpbf/osmformat.pb.go