I replaced gogo protobuf by the plain google protobuf implementation.

I had to rewrite some portions regarding nullability. Hope I got it right, there might be better way to accomplish.

Overall I was surprised that the CPU performance drop is below 3% , I was expecting much more.
The number of allocations is much higher for some benchmarks, since I had to alloc temporary objects because of nullability changes above. The only one I am concerned is the change in memory for BenchmarkLondon_relations-16 

```benchmark                                    old ns/op      new ns/op      delta
BenchmarkDiff_Marshal-16                     33708352       33497339       -0.63%
BenchmarkDiff_Unmarshal-16                   66501681       65426347       -1.62%
BenchmarkElementID_Sort-16                   648905         653604         +0.72%
BenchmarkFeatureIDs_Sort-16                  636472         631671         -0.75%
BenchmarkChange_MarshalXML-16                1790183        1744972        -2.53%
BenchmarkChangeSet_Marshal-16                79831          77896          -2.42%
BenchmarkUnmarshalWays-16                    5077           6401           +26.08%
BenchmarkUnmarshalRelations-16               5508           6938           +25.96%
BenchmarkChangeset_Marshal_gzip-16           450378         456373         +1.33%
BenchmarkChangeset_UnmarshalXML-16           3468850        3504131        +1.02%
BenchmarkUnmarshalChangeset-16               73306          81317          +10.93%
BenchmarkUnmarshalChangeset_gzip-16          137123         144377         +5.29%
BenchmarkRelations-16                        638028         633393         -0.73%
BenchmarkWay-16                              14600          14479          -0.83%
BenchmarkWays-16                             23030          22524          -2.20%
BenchmarkConvert-16                          1211337        1202652        -0.72%
BenchmarkConvertAnnotated-16                 947576         940853         -0.71%
BenchmarkConvert_NoID-16                     1084547        1063003        -1.99%
BenchmarkConvert_NoMeta-16                   996524         991362         -0.52%
BenchmarkConvert_NoRelationMembership-16     1157384        1148647        -0.75%
BenchmarkConvert_NoIDsMetaMembership-16      747993         748889         +0.12%
BenchmarkLondon-16                           257786472      262947900      +2.00%
BenchmarkLondon_withFiltersTrue-16           266369759      269555246      +1.20%
BenchmarkLondon_withFiltersFalse-16          208942287      206641137      -1.10%
BenchmarkLondon_nodes-16                     189334304      192258210      +1.54%
BenchmarkLondon_ways-16                      147202951      150169294      +2.02%
BenchmarkLondon_relations-16                 90644410       93034313       +2.64%
BenchmarkAndorra-16                          3074774010     3111769317     +1.20%

benchmark                                    old allocs     new allocs     delta
BenchmarkDiff_Marshal-16                     129484         129484         +0.00%
BenchmarkDiff_Unmarshal-16                   422916         422916         +0.00%
BenchmarkElementID_Sort-16                   1              1              +0.00%
BenchmarkFeatureIDs_Sort-16                  1              1              +0.00%
BenchmarkChange_MarshalXML-16                4385           4385           +0.00%
BenchmarkChangeSet_Marshal-16                291            242            -16.84%
BenchmarkUnmarshalWays-16                    74             97             +31.08%
BenchmarkUnmarshalRelations-16               81             105            +29.63%
BenchmarkChangeset_Marshal_gzip-16           315            265            -15.87%
BenchmarkChangeset_UnmarshalXML-16           23326          23326          +0.00%
BenchmarkUnmarshalChangeset-16               346            617            +78.32%
BenchmarkUnmarshalChangeset_gzip-16          390            660            +69.23%
BenchmarkRelations-16                        1417           1417           +0.00%
BenchmarkWay-16                              126            126            +0.00%
BenchmarkWays-16                             116            116            +0.00%
BenchmarkConvert-16                          10114          10114          +0.00%
BenchmarkConvertAnnotated-16                 10107          10107          +0.00%
BenchmarkConvert_NoID-16                     8501           8501           +0.00%
BenchmarkConvert_NoMeta-16                   6898           6898           +0.00%
BenchmarkConvert_NoRelationMembership-16     9531           9531           +0.00%
BenchmarkConvert_NoIDsMetaMembership-16      4702           4703           +0.02%
BenchmarkLondon-16                           4805957        4811479        +0.11%
BenchmarkLondon_withFiltersTrue-16           4805954        4811469        +0.11%
BenchmarkLondon_withFiltersFalse-16          1362152        1367677        +0.41%
BenchmarkLondon_nodes-16                     3392981        3398512        +0.16%
BenchmarkLondon_ways-16                      1793525        1799043        +0.31%
BenchmarkLondon_relations-16                 457585         463104         +1.21%
BenchmarkAndorra-16                          13038438       13038444       +0.00%

benchmark                                    old bytes     new bytes     delta
BenchmarkDiff_Marshal-16                     11232715      11232715      +0.00%
BenchmarkDiff_Unmarshal-16                   15975454      15975430      -0.00%
BenchmarkElementID_Sort-16                   24            24            +0.00%
BenchmarkFeatureIDs_Sort-16                  24            24            +0.00%
BenchmarkChange_MarshalXML-16                593113        593113        +0.00%
BenchmarkChangeSet_Marshal-16                68687         37545         -45.34%
BenchmarkUnmarshalWays-16                    6888          6984          +1.39%
BenchmarkUnmarshalRelations-16               7248          7536          +3.97%
BenchmarkChangeset_Marshal_gzip-16           891254        860106        -3.49%
BenchmarkChangeset_UnmarshalXML-16           1131472       1131467       -0.00%
BenchmarkUnmarshalChangeset-16               148593        173889        +17.02%
BenchmarkUnmarshalChangeset_gzip-16          225217        250481        +11.22%
BenchmarkRelations-16                        245465        245441        -0.01%
BenchmarkWay-16                              9622          9622          +0.00%
BenchmarkWays-16                             11640         11639         -0.01%
BenchmarkConvert-16                          913288        913285        -0.00%
BenchmarkConvertAnnotated-16                 830581        830577        -0.00%
BenchmarkConvert_NoID-16                     890772        890778        +0.00%
BenchmarkConvert_NoMeta-16                   702989        703007        +0.00%
BenchmarkConvert_NoRelationMembership-16     895883        895846        -0.00%
BenchmarkConvert_NoIDsMetaMembership-16      663290        663296        +0.00%
BenchmarkLondon-16                           918437913     938188674     +2.15%
BenchmarkLondon_withFiltersTrue-16           918437519     938187697     +2.15%
BenchmarkLondon_withFiltersFalse-16          360101121     379852203     +5.48%
BenchmarkLondon_nodes-16                     613038341     632790024     +3.22%
BenchmarkLondon_ways-16                      432007871     451758126     +4.57%
BenchmarkLondon_relations-16                 178275503     198026162     +11.08%
BenchmarkAndorra-16                          471254696     471255184     +0.00%```