@paulmach 2022-05-22
This looks good to me. I totally forgot about the marshal/unmarshal or nodes, ways, relations and changesets. Looks like I deprecated that 18 months ago https://github.com/paulmach/osm/commit/8fcda5dc49b4767df63eccb5a25f3e63d5b17f4d

That's where you see the biggest changes
```
BenchmarkUnmarshalWays-16                    5077           6401           +26.08%
BenchmarkUnmarshalRelations-16               5508           6938           +25.96%
BenchmarkUnmarshalChangeset-16               73306          81317          +10.93%
BenchmarkUnmarshalChangeset_gzip-16          137123         144377         +5.29%

benchmark                                    old allocs     new allocs     delta
BenchmarkUnmarshalWays-16                    74             97             +31.08%
BenchmarkUnmarshalRelations-16               81             105            +29.63%
BenchmarkUnmarshalChangeset-16               346            617            +78.32%
BenchmarkUnmarshalChangeset_gzip-16          390            660            +69.23%

benchmark                                    old bytes     new bytes     delta
BenchmarkUnmarshalWays-16                    6888          6984          +1.39%
BenchmarkUnmarshalRelations-16               7248          7536          +3.97%
BenchmarkUnmarshalChangeset-16               148593        173889        +17.02%
BenchmarkUnmarshalChangeset_gzip-16          225217        250481        +11.22%
```

I ended up moving that logic into my app a long time ago so I could iterate on the schema without worrying about backwards compatibility for other people.