@OlafFlebbeBosch 2022-05-23
Not using Marshal/Unmarshal  either -- in fact I only used osmpbf before.