@paulmach 2022-05-26
Thank you.

I published a new release with this change and will probably remove the node/way/relation encoding stuff in the next release

https://github.com/paulmach/osm/releases/tag/v0.4.0